Dev 0.0.11 创建数据库初始化方法（DataBaseInit::main()）

+++ src/main/java/io/mlqs/es/legal/db/MarriageLawRecordEntity.java
@@ -41,5 +41,5 @@ public class MarriageLawRecordEntity extends LegalRecordEntity{
      * 向量化
      */
     @Field(type = FieldType.Dense_Vector, dims = 1024)
-    private float[] vector;
+    private double[] vector;
 }
